more cleanup after merge
